+++ R/simulate.gllvm.R
@@ -70,7 +70,7 @@ simulate.gllvm = function (object, nsim = 1, seed = NULL, conditional = FALSE, .
     invPhis = matrix(rep(object$params$ZINB.inv.phi,each=nsim*nRows), ncol=nCols)
   if(object$family=="tweedie")
     phis = matrix(rep(object$params$phi, each = nsim*nRows), ncol = nCols)
-  if(object$family %in% c("gaussian", "gamma", "beta","ZIP","ZINB"))
+  if(object$family %in% c("gaussian", "gamma", "beta","ZIP","ZINB","ZIB"))
     phis = matrix(rep(object$params$phi, each = nsim*nRows), ncol = nCols)
   if(object$family == "ordinal"){
     if(object$zeta.struc=="species"){
@@ -111,6 +111,7 @@ simulate.gllvm = function (object, nsim = 1, seed = NULL, conditional = FALSE, .
                   "beta" = rbeta(nTot, shape1 = phis*prs, shape2 = phis*(1-prs)),
                   "ZIP" = ifelse(rbinom(nTot, size = 1, prob = phis) > 0, 0, rpois(nTot, lambda = prs)),
                   "ZINB" = ifelse(rbinom(nTot, size = 1, prob = phis) > 0, 0, rnbinom(nTot, size = invPhis, mu = prs)),
+                  "ZIB" = ifelse(rbinom(nTot, size = 1, prob = phis) > 0, 0, rbinom(nTot, size = rep(object$Ntrials,each=nsim*nRows), prob = prs)),
                   stop(gettextf("family '%s' not implemented ", object$family), domain = NA))
   # reformat as data frame with the appropriate labels
   newDat = as.data.frame(matrix(newDat,ncol=nCols))